requirements.txt
fastapi
uvicorn
motor
jinja2
python-dotenv
websockets